+++ app/src/main/java/com/yonasoft/jadedictionary/data/models/Word.kt
@@ -7,7 +7,7 @@ import androidx.room.PrimaryKey
 @Entity(tableName = "words")
 data class Word(
     @PrimaryKey(autoGenerate = true)
-    @ColumnInfo(name = "id") val id: Int? = null,
+    @ColumnInfo(name = "id") val id: Long? = null, // Changed from Int? to Long?
     @ColumnInfo(name = "simplified") val simplified: String? = null,
     @ColumnInfo(name = "traditional") val traditional: String? = null,
     @ColumnInfo(name = "pinyin") val pinyin: String? = null,
